parser: fix bug on right masked byte in hex string
The right mask was not tested as part of a full hex string, and
was not working. The issue was that the byte parser was used and
was using cut after parsing successfully a single digit.

Instead, do not use cut, and prefer trying to parse masked_byte before
byte.