feat: move true/false literals in primary_expression
There is a big issue with the current parsing, in that
both expression and primary_expression handle the parenthesis.
This leads to impossibilities to handle different inputs, such
as:
- (true) && false
- (1 + 1) * 2

To fix this, a rework will be done to remove primary_expression
as a 'special' parser used in many places, and only used as the
down-the-line parser only used by expression when all other
special syntaxes are handled.

First step is moving parsing of true & false in the primary
expression parser.