feat: return None in some wrong typing cases
When evaluating a rule, we might end up with the wrong types for a
given expression. In those cases, we want to generate an undefined
value.

This is fixed for some operations, and tested using a custom module
value that returns a value of a different type than the one declared and
used when type checking the rule.