feat: modify parsed file struct to keep order of imports
We need to keep the relative order between imports, includes and rules
in a Yara file, in order to properly resolve uses of rule names in
conditions.

The YaraFile object is thus transformed into a vec of an enum of
possible top-level AST values in the file.

The code in the boreal compiler is just modify to compile, but some
doc/API isn't great. This is intentional, this will be modify in the
following commits heavily.