feat: replace string value with bytes
String literals coming from the parser are guaranteed to be ascii, and
thus valid rust strings.

Module values however, are not so. Modules implemented in yara can
return byte strings without much of unicode validation. The string value
is thus changed in the evaluator to be a byte string, and just contain
arbitrary bytes.

As we can see, this does simplify a bit some module implementations.