feat: handle wide regex with word boundaries
The solution to handle this is... quite ugly. There isn't a very
pretty solution pre se, except maybe finding slices of the scanned mem
that are wide, unwiding them, and using them as input for wide regex.
But that is probably very bad performance wise...

The current solution is a bit ugly, but should work quite well.