feat: parse regex into AST, the convert it to rust regex exp
Previously, we gave the regex contents directly to the rust regex
engine. This causes many issues because the formats are not exactly
compatible.

Instead, we now parse the regex into a custom AST. Then, we convert this
AST into the best format for the rust regex crate. It could have been
its own AST type, but those depends on spans from parsing, and we do not
use the same spans. Instead, generate a string, which is easy to inspect
and test.