feat: use aho corasick to scan for literals
It is much faster to use AhoCorasick to scan for literals rather than
putting those in the RegexSet which is quite limited in size.
In addition, most yara rules uses strings instead of hex strings or
regexes.

This slows down a bit the execution for small number of rules as we do a
triple scan on the memory, but this can be optimized later on. It does
limit the performance explosion when using a lot of rules on big files.