feat: use a single aho corasick pass
Instead of using two AC passes, one case insensitive, one case
sensitive, use a single one.

Whether this one should be case insensitive or not is unclear. Benches
is unconclusive on which one is better, but the CI one leads a bit less
memory usage, so go with this one for the moment.