feat: verify hex string matches during ac pass
Instead of just storing the indication that there might be matches for
the hex strings, validate the matches during the AC pass.

This is far from trivial, as we need to reconstruct the regex
expression that express the hex string "before" and "after" the
literals.

This still needs some serious testing.