feat: limit size of literals used in AC pass
Pick atoms of size max 4 out of the literals extracted for the AC pass.

This serves two purposes:
- Make the AC smaller
- Use higher quality literals in the search

This gives around the same performances, but drastically reduces the
memory usage, giving us the opportunity of using the dfa version of the
AC algo. This dfa version is much more performant.