feat: rework the undefined kind from an Option to a Result
To properly compute rule's results, we need to distinguish the two
undefined or poison kinds, So that for example `uint8(0)` on an empty
slice does not lead to a rule's result being undecidable.

To do so, the Option used during evaluation is transformed into a
Result, which allows rethrowing as usual, but gives multiple states for
the thrown value.