feat: add module names option in boreal-cli
The particularity is that this option acts as its own mode: if given, it
will print the names and exit, so the required arguments (rules & input)
are not required if this option is specified.

This was not trivial to find how to express this with clap, but the
required_unless_present seems to work.