feat: compare module values with libyara
Retrieve the module values from libyara, and compare with our own
values. This replaces the very verbose coverage tests, and will ensure
there are no typos, and no values is forgotten.

This new helper is used in the macho coverage tests, and will be used in
the other modules as well in future commits.