feat: do not modify AcResult if multiple matches is empty
The AcmatchStatus::Multiple is always returned if we reach the reverse
validator passes, however it can return an empty vec. In that case do
not save the ac result as matches. This has no impact on the overall
results, but may improve stats or inspection of the AcResult enum to
make it less confusing.