feat: rework the literals extraction
Rework the literals extraction algorithm to improve it and handle
expanding classes to generate better literals.

The main goal of this commit is to handle expanding classes if it allows
generating atoms with much better quality (or really, longer length).

Mainly, it is to be able to expand badly optimized hex strings so that
we can get atoms from them. For example:

- `{ AB F? 01 }`: we now generate 16 literals of length 3, better
   than a single literal of length 1
- `{ AB ?? 00 00 }`: we now generate 256 literals of length 4,
   better than a single literal of length 1

This is made more generic than just masked bytes, by expanding all
possible classes.

In order to keep the complexity of the code low, there is also a change
of behavior: literals from alternations are no longer concatenated to
literals outside of those alternations. For example, while we previously
handled `a(b|c)` by generating the literals `ab` and `ac`, we now either
use `a` or use `b` and `c`.

This seems to be in practice completely ok: out of all the rules that
are used for benchmarking and testing, extremely few of them had worse
atoms because of this change.