feat: improve pe module values tests wrt rich_signature
YARA always set a pe.rich_signature object, with two functions that
will always return undefined.

Boreal does not work the same way, and does not set this object if a
rich signature is not present.

This difference means we have to ignore this diff when comparing module
values, which is annoying. With the addition of tests with process
memory, this brings even more such differences (because for a dll, yara
will either report this object or not (on process memory), so do a
little change to remove this difference and help tests.