feat: update FragmentedMemory::fetch to be failable
Make it possible to return None when fetching memory. This could have
been a Result as well, but the error would have just been ignored.

Another possiblity I played with was returning two kinds of errors,
either recoverable (we keep scanning) or unrecoverable (we abort the
scan). This is however a bit shaky on the semantics, since during
evaluation, we may fetch regions, but we have no way of bubbling up the
unrecoverable error in that case.

So for the moment, just use an Option. If a more complex usecase arises
in the future, this could change.