feat: handle access to contiguous fragmented regions
For some expressions, we want to be able to make them work even if
regions are chunked. To do so, we need a new API to iterate on multiple
distinct byte slices. This is done with the new Memory::on_range method.
This uses a callback instead of returning an iterator to guarantee that
only one region can be fetched at a given time, which is a constraint
used in the memory trait.