exclude test assets in package
This PR excludes the test assets from the published package for the crate.  As is, a number of AVs trigger on the contents of the resulting package, which impacts mirrors of crates.io.

As the test & assets are used by crate contributors, but not crate consumers, I believe this should not have any downsides.