ci: do not use -Dwarnings rust flag on coverage
The coverage job uses the nightly toolchain due to a dependency on a
nightly feature. Since all jobs uses the -Dwarnings rust flag however,
this meant this job can break every day as the nightly toolchain adds
new warnings.

To avoid this issue and have a more stable job, stop setting this flag
for all jobs, and only do it for jobs where it makes sense.