fix: fix bug in global rules declared after normal rules
When a global rule was declared before a normal rule, and both declared
strings, the matches of the strings of the normal rule were used to
evaluate the global rule.

This is fixed by splicing the variables of the global rules earlier in
the variables vec, so that the whole variables vec respects the
global_rules ++ rules order.