feat: add CompilerBuilder object
Add a builder for the compiler object. This builder is only used
for the moment to add modules, but more will be added.

This builder solves an issue when adding modules, in that existing
modules could not be replaced by the one added, which made for a much
worse API and the impossibility of overwriting a module configuration
with a new one.

This is now solved as with a CompilerBuilder, it is guaranteed that
added modules have not yet been used by any rules, so they can be
replaced.

This makes creating a compiler a bit more complex when custom modules
are added, but this is not a huge change.