feat: add compiler profile to prioritize memory usage
Add a compiler profile option to CompilerBuilder. This will allow
picking between:

- speed: prioritize scanning speed (current impl & default)
- memory: prioritize memory usage
- automatic: pick between the two depending on the size of the rules.

The memory profile is only implemented for the aho-corasick for the
moment. Changing the regex engine used for validators could be an
option, however it isn't trivial to do, and we would make implementing
an "automatic" profile harder, since "automatic" only works if
compilation is done when converting the compiler to the scanner.