feat[boreal-cli]: add ignored arguments in yr subcommand
In order to allow the boreal yr subcommand to be a substitute
of the yara CLI, lets accepts some arguments which make no
sense here but are ok to ignored.

The last argument that is not handled is the atom quality
table, but it is preferable to not accept it for the moment
to avoid tricking the user into thinking this quality table
is properly loaded.