refacto/boreal-parser: move recursion limit into Input object
Instead of harcoding recursion limits, make those variables
used during parsing. Unfortunately, it is not easy to propagate
data during parsing due to how it is written, with function combinators.

To circumvent this issue, those limits are put directly in the "Input"
object, which is a bit ugly, but works well. This only adds two u8
values in it which due to existing padding does not end up growing the
object at all.