feat: improve alternations handling for atoms extraction
This MR brings a much needed improvement in the atoms extraction
logic so that several rules used in the wild that had performance
issues can be properly handled.

The main idea of the change is to be able to join alternates to
surrounding bytes while allowing the alternates to contain jumps.

The idea is to split alternates into an run open on the left, and
a run open on the right. For example, for this regex:

abc.def.ghi

- the "abc" part is open on the left
- the "def" part is closed in the middle
- the "ghi" part is open on the right

Lets see this in an alternations:

- a(1.2.3|4.5.6)z

This can be split as follows:

- <prefix>(<pre1>.<mid1>.<post1>|<pre2>.<mid2>.<post2>)<suffix>

There are three possible atoms from this:

- <prefix><pre> (ie ["a1", "a4"] in the example)
- <mid> (ie ["2", "5"] in the example)
- <post><suffix> (ie ["3z", "6z"] in the example)

This is simple in theory, but not that trivial to handle, especially
as the regex must also be split into a reverse and a forward validator
from the position of the extracted literals.

One illustration of this is for example that the "<mid>" extraction
is actually impossible: since there is nothing linking the reverse and
forward part from the literal, false positives could be generated.
One such example would be the regex "1.2.3|4.2.5", and the extracted
literal "2".

To properly handle the generation of the reverse and forward validator,
the position of the literals inside the alternation must be reused
properly to split the alternation at the right places for the validators
to function. This requires a rework of this logic which makes it a
bit more complex, but this stays fairly manageable.