feat: add PeData object to allows overriding pe.is_signed
Add a way to provide a value for pe.is_signed. This allows
calling WinVerifyTrust (or any other source of validation)
first, and then exposing the result in the `pe.is_signed`
flag.

This gives an alternative to the authenticode-verify feature
and a way to have a proper value for this field.